Makefile for MOVE and PLACE programs, plus their DSO's.

+++ insts/std/MOVE/BASE.cpp
@@ -70,7 +70,7 @@ BASE::~BASE()
    }
 }
 
-int BASE::init(float p[], short n_args)
+int BASE::init(float p[], int n_args)
 {
     int    flag, UseMikes, cartflag = 0;
     float  outskip, inskip, abs_factor, rvb_time;
