+++ insts/std/MOVE/MOVE.cpp
@@ -71,7 +71,7 @@ MOVE::~MOVE()
     delete [] rholoc;
 }
 
-int MOVE::localInit(float *p, short n_args)
+int MOVE::localInit(float *p, int n_args)
 {
     if (n_args < 6)
         die(name(), "Wrong number of args.");
