+++ insts/std/MOVE/PLACE.h
@@ -11,7 +11,7 @@ public:
     virtual ~PLACE();
 protected:
     virtual const char * name() { return "PLACE"; }
-    virtual int localInit(float *, short);
+    virtual int localInit(float *, int);
     virtual int finishInit(double, double *);
     virtual int updatePosition(int);
     virtual void get_tap(int, int, int, int);
